Add wait condition to Unistaller pre-delete hook
Pre-delete hook now waits that all peering are correctly deactivated before exiting. This prevent the possibility of having resources remaining on the cluster after uninstall.
